+++ com.unity.toonshader/Editor/ToonGUI-RenderingPerChannel.cs
@@ -46,7 +46,7 @@ namespace UnityEditor.Rendering.Toon
         internal  void RenderingPerChennelsSettingHDRP(Material material)
         {
 
-            _PerChanelShaderSettings_Foldout = Foldout(_PerChanelShaderSettings_Foldout, "【Mask Rendering Settings】");
+            _PerChanelShaderSettings_Foldout = Foldout(_PerChanelShaderSettings_Foldout, "Mask Rendering Settings");
             if (!_PerChanelShaderSettings_Foldout)
             {
                 return;