+++ com.unity.toonshader/Editor/UTS3GUI.cs
@@ -969,16 +969,16 @@ namespace UnityEditor.Rendering.Toon
             public static GUIContent baseColorText = new GUIContent("Base Map", "Base Color : Texture(sRGB) × Color(RGB) Default:White");
             public static GUIContent firstShadeColorText = new GUIContent("1st Shading Map", "1st ShadeColor : Texture(sRGB) × Color(RGB) Default:White");
             public static GUIContent secondShadeColorText = new GUIContent("2nd Shading Map", "2nd ShadeColor : Texture(sRGB) × Color(RGB) Default:White");
-            public static GUIContent normalMapText = new GUIContent("NormalMap", "NormalMap : Texture(bump)");
-            public static GUIContent highColorText = new GUIContent("HighColor", "High Color : Texture(sRGB) × Color(RGB) Default:Black");
-            public static GUIContent highColorMaskText = new GUIContent("HighColor Mask", "HighColor Mask : Texture(linear)");
-            public static GUIContent rimLightMaskText = new GUIContent("RimLight Mask", "RimLight Mask : Texture(linear)");
+            public static GUIContent normalMapText = new GUIContent("Normal Map", "Normal Map : Texture(bump)");
+            public static GUIContent highColorText = new GUIContent("High Light", "High Light : Texture(sRGB) × Color(RGB) Default:Black");
+            public static GUIContent highColorMaskText = new GUIContent("High Light Mask", "High Light Mask : Texture(linear)");
+            public static GUIContent rimLightMaskText = new GUIContent("Rim Light Mask", "RimLight Mask : Texture(linear)");
             public static GUIContent matCapSamplerText = new GUIContent("MatCap Sampler", "MatCap Sampler : Texture(sRGB) × Color(RGB) Default:White");
             public static GUIContent matCapMaskText = new GUIContent("MatCap Mask", "MatCap Mask : Texture(linear)");
-            public static GUIContent angelRingText = new GUIContent("AngelRing", "AngelRing : Texture(sRGB) × Color(RGB) Default:Black");
-            public static GUIContent emissiveTexText = new GUIContent("Emissive", "Emissive : Texture(sRGB)× EmissiveMask(alpha) × Color(HDR) Default:Black");
+            public static GUIContent angelRingText = new GUIContent("Angel Ring", "AngelRing : Texture(sRGB) × Color(RGB) Default:Black");
+            public static GUIContent emissiveTexText = new GUIContent("Emissive Map", "Emission : Texture(sRGB)× EmissiveMask(alpha) × Color(HDR) Default:Black");
             public static GUIContent shadingGradeMapText = new GUIContent("Shading Grade Map", "Specify shadow-prone areas in UV coordinates. Shading Grade Map : Texture(linear)");
-            public static GUIContent firstPositionMapText = new GUIContent("Specify the position of fixed shadows that fall in 1st shade color areas in UV coordinates. 1st Position Map : Texture(linear)");
+            public static GUIContent firstPositionMapText = new GUIContent("1st Shade Position Map", "Specify the position of fixed shadows that fall in 1st shade color areas in UV coordinates. 1st Position Map : Texture(linear)");
             public static GUIContent secondPositionMapText = new GUIContent("2nd Shade Position Map", "Specify the position of fixed shadows that fall in 2nd shade color areas in UV coordinates. 2nd Position Map : Texture(linear)");
             public static GUIContent outlineSamplerText = new GUIContent("Outline Sampler", "Outline Sampler : Texture(linear)");
             public static GUIContent outlineTexText = new GUIContent("Outline tex", "Outline Tex : Texture(sRGB) Default:White");
@@ -1147,7 +1147,7 @@ namespace UnityEditor.Rendering.Toon
 
             EditorGUILayout.Space();
 
-            _HighColor_Foldout = Foldout(_HighColor_Foldout, "High Color Settings");
+            _HighColor_Foldout = Foldout(_HighColor_Foldout, "High Light Settings");
             if (_HighColor_Foldout)
             {
                 EditorGUI.indentLevel++;
@@ -1861,7 +1861,7 @@ namespace UnityEditor.Rendering.Toon
                 EditorGUILayout.EndHorizontal();
 #else
                 GUI_Toggle(material, "3 Basic Colors", ShaderPropIs_NormalMapToBase, MaterialGetInt(material, ShaderPropIs_NormalMapToBase) != 0);
-                GUI_Toggle(material, "High Color", ShaderPropNormalMapToHighColor, MaterialGetInt(material, ShaderPropNormalMapToHighColor) != 0);
+                GUI_Toggle(material, "High Light", ShaderPropNormalMapToHighColor, MaterialGetInt(material, ShaderPropNormalMapToHighColor) != 0);
                 GUI_Toggle(material, "Rim Light", ShaderPropIsNormalMapToRimLight, MaterialGetInt(material, ShaderPropIsNormalMapToRimLight) != 0);
 
 #endif
@@ -1909,6 +1909,7 @@ namespace UnityEditor.Rendering.Toon
 
                 if (material.HasProperty("_StepOffset"))//Items not in Mobile & Light Mode.                
                 {
+#if false
                     //Line();
                     //EditorGUILayout.Space();
                     _AdditionalLookdevs_Foldout = FoldoutSubMenu(_AdditionalLookdevs_Foldout, "● Additional Settings");
@@ -1916,6 +1917,13 @@ namespace UnityEditor.Rendering.Toon
                     {
                         GUI_AdditionalLookdevs(material);
                     }
+#else
+                    _AdditionalLookdevs_Foldout = FoldoutSubMenu(_AdditionalLookdevs_Foldout, "Point Light Settings");
+                    if (_AdditionalLookdevs_Foldout)
+                    {
+                        GUI_AdditionalLookdevs(material);
+                    }
+#endif
                 }
             }
         }
@@ -1950,7 +1958,7 @@ namespace UnityEditor.Rendering.Toon
             EditorGUI.BeginDisabledGroup(!isEnabled);
             {
                 EditorGUI.indentLevel++;
-                m_MaterialEditor.RangeProperty(tweak_SystemShadowsLevel, "System Shadows Level");
+                m_MaterialEditor.RangeProperty(tweak_SystemShadowsLevel, "System Shadow Level");
                 GUI_SetRTHS(material);
                 EditorGUI.indentLevel--;
                 EditorGUILayout.Space();
@@ -1966,10 +1974,10 @@ namespace UnityEditor.Rendering.Toon
                 if (material.GetInt(ShaderPropUtsTechniqe) == (int)_UTS_Technique.DoubleShadeWithFeather)   //DWF
                 {
                     GUILayout.Label("Mode: Double Shade With Feather", EditorStyles.boldLabel);
-                    m_MaterialEditor.RangeProperty(baseColor_Step, "BaseColor Step");
+                    m_MaterialEditor.RangeProperty(baseColor_Step, "Base Color Step");
                     m_MaterialEditor.RangeProperty(baseShade_Feather, "Base/Shade Feather");
-                    m_MaterialEditor.RangeProperty(shadeColor_Step, "ShadeColor Step");
-                    m_MaterialEditor.RangeProperty(first2nd_Shades_Feather, "1st/2nd_Shades Feather");
+                    m_MaterialEditor.RangeProperty(shadeColor_Step, "Shade Color Step");
+                    m_MaterialEditor.RangeProperty(first2nd_Shades_Feather, "1st/2nd Shade Feather");
                     //Sharing variables with ShadingGradeMap method.
 
                     material.SetFloat(ShaderProp1st_ShadeColor_Step, material.GetFloat(ShaderPropBaseColor_Step));
@@ -1980,10 +1988,10 @@ namespace UnityEditor.Rendering.Toon
                 else if (material.GetInt(ShaderPropUtsTechniqe) == (int)_UTS_Technique.ShadingGradeMap)
                 {    //SGM
                     GUILayout.Label("Mode: Shading Grade Map", EditorStyles.boldLabel);
-                    m_MaterialEditor.RangeProperty(first_ShadeColor_Step, "1st ShaderColor Step");
-                    m_MaterialEditor.RangeProperty(first_ShadeColor_Feather, "1st ShadeColor Feather");
-                    m_MaterialEditor.RangeProperty(second_ShadeColor_Step, "2nd ShadeColor Step");
-                    m_MaterialEditor.RangeProperty(second_ShadeColor_Feather, "2nd ShadeColor Feather");
+                    m_MaterialEditor.RangeProperty(first_ShadeColor_Step, "1st Shade Color Step");
+                    m_MaterialEditor.RangeProperty(first_ShadeColor_Feather, "1st Shade Color Feather");
+                    m_MaterialEditor.RangeProperty(second_ShadeColor_Step, "2nd Shade Color Step");
+                    m_MaterialEditor.RangeProperty(second_ShadeColor_Feather, "2nd Shade Color Feather");
                     //Share variables with DoubleWithFeather method.
                     material.SetFloat(ShaderPropBaseColor_Step, material.GetFloat(ShaderProp1st_ShadeColor_Step));
                     material.SetFloat(ShaderPropBaseShade_Feather, material.GetFloat(ShaderProp1st_ShadeColor_Feather));
@@ -2001,9 +2009,9 @@ namespace UnityEditor.Rendering.Toon
 
         void GUI_AdditionalLookdevs(Material material)
         {
-            GUILayout.Label("    Settings for PointLights in ForwardAdd Pass");
+//            GUILayout.Label("    Point Light Settings");
             EditorGUI.indentLevel++;
-            m_MaterialEditor.RangeProperty(stepOffset, "Step Offset for PointLights");
+            m_MaterialEditor.RangeProperty(stepOffset, "Step Offset");
 #if USE_TOGGLE_BUTTONS
 
             EditorGUILayout.BeginHorizontal();
@@ -2025,7 +2033,7 @@ namespace UnityEditor.Rendering.Toon
             }
             EditorGUILayout.EndHorizontal();
 #else
-            GUI_Toggle(material, "PointLights Hi-Cut Filter", ShaderPropIsFilterHiCutPointLightColor, MaterialGetInt(material, ShaderPropIsFilterHiCutPointLightColor) != 0);
+            GUI_Toggle(material, "Point Light Hi-Cut Filter", ShaderPropIsFilterHiCutPointLightColor, MaterialGetInt(material, ShaderPropIsFilterHiCutPointLightColor) != 0);
 #endif
             EditorGUI.indentLevel--;
             EditorGUILayout.Space();
@@ -2035,7 +2043,7 @@ namespace UnityEditor.Rendering.Toon
         void GUI_HighColor(Material material)
         {
             m_MaterialEditor.TexturePropertySingleLine(Styles.highColorText, highColor_Tex, highColor);
-            m_MaterialEditor.RangeProperty(highColor_Power, "HighColor Power");
+            m_MaterialEditor.RangeProperty(highColor_Power, "High Light Power");
 
             if (!_SimpleUI)
             {
@@ -2157,11 +2165,11 @@ namespace UnityEditor.Rendering.Toon
                     EditorGUI.indentLevel--;
                 }
 #else
-                var ret = GUI_Toggle(material, "Shadow Mask on High Color", ShaderPropIs_UseTweakHighColorOnShadow, MaterialGetInt(material, ShaderPropIs_UseTweakHighColorOnShadow) != 0);
+                var ret = GUI_Toggle(material, "Shadow Mask on High Lights", ShaderPropIs_UseTweakHighColorOnShadow, MaterialGetInt(material, ShaderPropIs_UseTweakHighColorOnShadow) != 0);
                 EditorGUI.BeginDisabledGroup(!ret);
                 {
                     EditorGUI.indentLevel++;
-                    m_MaterialEditor.RangeProperty(tweakHighColorOnShadow, "High Color Power on Shadow");
+                    m_MaterialEditor.RangeProperty(tweakHighColorOnShadow, "High Light Power on Shadows");
                     EditorGUI.indentLevel--;
                 }
                 EditorGUI.EndDisabledGroup();
@@ -2175,10 +2183,10 @@ namespace UnityEditor.Rendering.Toon
             //Line();
             //EditorGUILayout.Space();
 
-            GUILayout.Label("    High Color Mask", EditorStyles.boldLabel);
+            GUILayout.Label("    High Light Mask", EditorStyles.boldLabel);
             EditorGUI.indentLevel++;
             m_MaterialEditor.TexturePropertySingleLine(Styles.highColorMaskText, set_HighColorMask);
-            m_MaterialEditor.RangeProperty(tweak_HighColorMaskLevel, "High Color Mask Level");
+            m_MaterialEditor.RangeProperty(tweak_HighColorMaskLevel, "High Lgiht Mask Level");
             EditorGUI.indentLevel--;
 
             EditorGUILayout.Space();
@@ -3806,7 +3814,7 @@ namespace UnityEditor.Rendering.Toon
             m_MaterialEditor.RangeProperty(unlit_Intensity, "Unlit Intensity");
             RestoreGUIColor();
             EditorGUILayout.BeginHorizontal();
-            EditorGUILayout.PrefixLabel("SceneLights Hi-Cut Filter");
+            EditorGUILayout.PrefixLabel("Scene Light Hi-Cut Filter");
             //GUILayout.Space(60);
             if (material.GetFloat(ShaderPropIs_Filter_LightColor) == 0)
             {
@@ -4039,7 +4047,7 @@ namespace UnityEditor.Rendering.Toon
             GUI_Toggle(material, "Base Color", ShaderPropIsLightColor_Base, MaterialGetInt(material, ShaderPropIsLightColor_Base)!= 0);
             GUI_Toggle(material, "1st Shading Color", ShaderPropIs_LightColor_1st_Shade, MaterialGetInt(material, ShaderPropIs_LightColor_1st_Shade) != 0);
             GUI_Toggle(material, "2nd Shading Color", ShaderPropIs_LightColor_2nd_Shade, MaterialGetInt(material, ShaderPropIs_LightColor_2nd_Shade) != 0);
-            GUI_Toggle(material, "High Color", ShaderPropIs_LightColor_HighColor, MaterialGetInt(material, ShaderPropIs_LightColor_HighColor) != 0);
+            GUI_Toggle(material, "High Light", ShaderPropIs_LightColor_HighColor, MaterialGetInt(material, ShaderPropIs_LightColor_HighColor) != 0);
             GUI_Toggle(material, "Rim Light", ShaderPropIs_LightColor_RimLight, MaterialGetInt(material, ShaderPropIs_LightColor_RimLight) != 0);
             GUI_Toggle(material, "Ap_RimLight", ShaderPropIs_LightColor_Ap_RimLight, MaterialGetInt(material, ShaderPropIs_LightColor_Ap_RimLight) != 0);
             GUI_Toggle(material, "MatCap", ShaderPropIs_LightColor_MatCap, MaterialGetInt(material, ShaderPropIs_LightColor_MatCap) != 0);
@@ -4128,7 +4136,7 @@ namespace UnityEditor.Rendering.Toon
 #else
             EditorGUI.BeginChangeCheck();
             var prop = ShaderPropIs_Filter_LightColor;
-            var label = "SceneLights Hi-Cut Filter";
+            var label = "Scene Light Hi-Cut Filter";
             var value = MaterialGetInt(material, prop);
             var ret = EditorGUILayout.Toggle(label, value != 0);
             if (EditorGUI.EndChangeCheck())